feat: expand metadata returned by search call, allow filter on collection name
Adds the collection name and the identifiers of any attestations
contained within the collection to the response from GetBySubjectDigest.

Adds a parameter to limit results by a collection's name.

Adds more options to archivistctl to interact with archivist.

Some code re-org to split up metadata stores from object stores, and
moves the object store calls out of the mysqlstore.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>