bug: signature column cannot handle all signatures (#26)
By default ent generates a mysql column of type `varchar(255)` for
string types. This poses a problem for RSA-PSS signatures as these
signatures have lengths equal to the size of the RSA key.

For instance an RSA key with length of 4096bits will generate a
signature with length 4096bits, or 512 bytes/characters. This obviously
will not fit in a `varchar(255)` column.

This PR changes the signature column to use a MySQL `text` column type
which can handle strings of up to 65,535 bytes.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>