fix: add Time type definition for gqlgen
Gqlgen gets angry because it doesn't recognize the Time datatype. The
solution is to define a separate graphql schema with the Time type
defined before it generates the rest of our schema for us.

See https://github.com/99designs/gqlgen/issues/1463

We already had gqlgen configured to look for an archivista.graphql file
for this reason (see gqlgen.yaml), but apparently we did not check it
in.