Auto-update pre-commit hooks PR
Using pre-commit can be very helpful for new contributors, but it can
also bring some difficulties.

Sometimes, a contributor will have a different version of the hook,
which is not part of his PR. We should keep it updated upstream to
avoid different checks due to mismatched versions.

This commit adds a GHA workflow to submit updates (like dependabot)

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>