chore: simplify development routine using container
This commit simplifies the development routing using containers and
shortcuts using Makefile.

The user can start the development environment using the command
`make run-dev` command, which will create and start all the
containers required for Archivista. The significant advantage is that
the user wants to test the container, so it is not necessary to build
Archivista, rebuild the container, and run it. It is done
automatically by CompileDaemon.

This command also includes:
`make stop` to stop all container
`make clean` to clean all Archivista containers

Added a shortcut to run the Archivista tests
`make tests`

This strategy is inspired by
- Warehouse development: https://warehouse.pypa.io/development/getting-started.html#running-the-warehouse-container-and-services
- RSTUF development: https://github.com/repository-service-tuf/repository-service-tuf-api#running-api-development

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>