refactoring: testifysec/archivista-api as a pkg
The current "github.com/testifysec/archivista-api" was detached from
the Archivista from the historical strategy by TestifySec.

TestifySec donated the Archivista project to the in-toto project,
which makes more sense to the API package being part of the
Archivista source code.

This commit adds "github.com/testifysec/archivista-api" as pkg in
Archivista.

Closes #113

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>