Initial improvement for Archivista contributors (#114)
* refactor: compose name standard

This commit changes `compose.psql.yml` to `compose-psql.yml` as the
double extension can be confusing for some tools.

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>

* chore: simplify development routine using container

This commit simplifies the development routing using containers and
shortcuts using Makefile.

The user can start the development environment using the command
`make run-dev` command, which will create and start all the
containers required for Archivista. The significant advantage is that
the user wants to test the container, so it is not necessary to build
Archivista, rebuild the container, and run it. It is done
automatically by CompileDaemon.

This command also includes:
`make stop` to stop all container
`make clean` to clean all Archivista containers

Added a shortcut to run the Archivista tests
`make tests`

This strategy is inspired by
- Warehouse development: https://warehouse.pypa.io/development/getting-started.html#running-the-warehouse-container-and-services
- RSTUF development: https://github.com/repository-service-tuf/repository-service-tuf-api#running-api-development

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>

* docs: Add contributing documentation to Archivista

This commit includes the CONTRIBUTING.md seed.
The initial content shares how to start contributing, the
requirements, and how to quick-start the development environment.

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>

* changes / suggested improvements on CONTRIBUTING.md

---------

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>
Co-authored-by: Tom Meadows <tom@tmlabs.co.uk>