tests: Include UT tests to pkg/api/download
Add unit tests for `pkg/api/download` coverage as 85%

The UT found a bug in the `DownloadWithWriter`. This bug returns nil
instead of error and affects users while downloading envelopes. For
example, the `archivistactl retrieve envelope` command, if giving an
invalid server or if the server is unavailable, will return no error.

In the `archivistactl`, the user will behave the same when the
server is available, and there is no envelope with the expected
`gitoid`.

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>