test: consolidate tests and add code coverage
This commit consolidates the `tests` and `e2e` tests in one.
The command `make tests` already run all the tests (unit and e2e).

This commit also includes in witness.yml the action to save the
cover.out to be used later by the pipeline to download the cover.out and
upload to the CodeCov.

Signed-off-by: Kairo de Araujo <kairo.araujo@testifysec.com>