fix: consistency protection storing attestations
COMMIT MESSAGE

Adds protection to the users while storing attestations

As Archivista relies on Object Storage/Filesystem to store the blob
attestation and the SQL server to register the attestations for
querying using GraphSQL, the flow needs to be consistent, as the
services can fail.

Ideally, the Store should happen transactional and not finish with
inconsistency: file available in the SQL but not in the metadata
Storage, for example. If it happens, the user will query the SQL
but will not be able to retrieve/download the attestation blob.

A minor fix is done in this PR, first adding the to the metadata
storage and after registering in the SQL server.
So, if the metadata storage fails, it will not continue and save
it in the SQL server.

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>