refactoring: rename store APIs to upload
This refactoring focuses on the code readability and maintainability
easier for new contributors.

`Store` -> `Upload`
`StoreWithHeaders` -> `UploadWithHeaders`

The `Store` will be exclusive for the Store method used by
`metadataStore` and `objectStore`, causing less confusion.

It renames the functions but keeps backward compatibility to allow
the current users the possibility to migrate.

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>