fix: Archivista Download API responses
Archivista Download API returned 500 if a GitOID is available while
the user calls the `/v1/download/<gitoid>` (`/download/<gitoid>`)
endpoint.

It is not an Internal Server Error (500) but a not-found for the
GitOID endpoint.

This PR fixes the HTTP status code, giving the users a more precise
response.

This commit also includes a better list of return codes to the
Swagger documentation.

It includes the UT

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>